requirements.txt
asgiref==3.8.1
biopython==1.83
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
dj-database-url==2.2.0
Django==5.0.6
fasteners==0.19
fonttools==4.53.0
GridDataFormats==1.0.2
gunicorn==22.0.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
mda-xdrlib==0.2.0
MDAnalysis==2.7.0
mmtf-python==1.1.3
mrcfile==1.5.0
msgpack==1.0.8
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
#pillow==10.3.0
plotly==5.22.0
prolif==2.0.3
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
rcsbsearchapi==1.5.1
rdkit==2023.9.4
requests==2.32.3
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
xlsxwriter==3.2.0
beautifulsoup4==4.12.3
djangorestframework==3.15.2
django-rest-swagger==2.2.0
drf-spectacular==0.27.2
IPython==8.25.0
scikit-learn
umap-learn
django-plotly-dash
seaborn